jwt-flask/requirements.txt
flask
flask-sqlalchemy
pyjwt
cryptography
python-dotenv